requirements.txt
# Requirements: OSX:libjpeg, Ubuntu14.04:libjpeg-dev, python-dev

Django==1.9.4
django-filter==0.13.0
django-import-export==0.4.5
django-rest-swagger==0.3.5
djangorestframework==3.3.3
django-cors-headers==1.1.0
Markdown==2.6.6
PyYAML==3.11
whitenoise==2.0.6
gunicorn==19.4.5
#psycopg2==2.6.1
whitenoise==2.0.6
dj-database-url==0.4.0

#Pillow==3.1.1
#djoser==0.4.3
#django-constance[database]==1.2
#django-picklefield==0.3.2
